Fixed volume bug in AddSound - volume parameter was specified as an int, preventing fractional volume values

+++ src/sbs/elevator.cpp
@@ -3736,7 +3736,7 @@ void Elevator::SetMessageSound(bool direction, const char *filename)
 	UseMessageSounds = true;
 }
 
-Object* Elevator::AddSound(const char *name, const char *filename, Ogre::Vector3 position, bool loop, int volume, int speed, float min_distance, float max_distance, float doppler_level, float cone_inside_angle, float cone_outside_angle, float cone_outside_volume, Ogre::Vector3 direction)
+Object* Elevator::AddSound(const char *name, const char *filename, Ogre::Vector3 position, bool loop, float volume, int speed, float min_distance, float max_distance, float doppler_level, float cone_inside_angle, float cone_outside_angle, float cone_outside_volume, Ogre::Vector3 direction)
 {
 	//create a looping sound object
 	Sound *sound = new Sound(object, name, false);
